MM_30482 : Renamed entity_factory into operators
